move exponential backoff to lib
